Add support for abilities from taxonomy as well as ability parameters.
As part of this we also update the state generation process. We want the states to be added immediately onto the object as they are generated (instead of being cached in a temp dict inside the factory function) and we want all object metadata links to be ready at the time of state creation for assertion purposes.